+++ src/components/utils/roleApi.js
@@ -1,9 +1,17 @@
 import { api, getHeader, getToken, getHeaderForFormData } from "./AppFunction";
 import moment from "moment";
 
-export async function createRole(name, users){
+const isRoleTypeValid = ["SUPER_ADMIN", "ADMIN", "STORAGE_FOREMAN", "STORAGE_EMPLOYEE", "STORAGE_MANAGER"];
+
+export async function createRole(name, users, roleTypes, permissionIds){
     try{
-        const requestBody = {name, users};
+        if(!name || typeof name !== "string" || name.trim() ==="" ||
+            !Array.isArray(users) || users.length === 0 ||
+            !isRoleTypeValid.includes(roleTypes?.toUpperCase()) ||
+            !(permissionIds instanceof Set) || permissionIds.size === 0){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
+        const requestBody = {name, users,roleTypes, permissionIds};
         const response = await api.post(`${import.meta.env.VITE_API_BASE_URL}/roles/create-new-role`,requestBody,{
             headers:getHeader()
         });
@@ -19,9 +27,17 @@ export async function createRole(name, users){
     }
 }
 
-export async function updateRole(roleId, name, users){
+export async function updateRole(roleId, name, users, roleTypes, permissionIds){
     try{
-        const requestBody = {name, users};
+        if(
+            !roleId ||
+            !name || typeof name !== "string" || name.trim() ==="" ||
+            !Array.isArray(users) || users.length === 0 ||
+            !isRoleTypeValid.includes(roleTypes?.toUpperCase()) ||
+            !(permissionIds instanceof Set) || permissionIds.size === 0){
+            throw new Error("Sva polja moraju biti validna i popunjena");
+        }
+        const requestBody = {name, users,roleTypes, permissionIds};
         const response = await api.put(`${import.meta.env.VITE_API_BASE_URL}/roles/update/${roleId}`,requestBody,{
             headers:getHeader()
         });
@@ -39,6 +55,9 @@ export async function updateRole(roleId, name, users){
 
 export async function deleteRole(roleId){
     try{    
+        if(!roleId){
+            throw new Error("Dati ID za role nije pronadjen");
+        }
         const response = await api.delete(`${import.meta.env.VITE_API_BASE_URL}/roles/delete/${roleId}`,{
             headers:getHeader()
         });
@@ -61,6 +80,9 @@ export async function getAllRoles(){
 
 export async function getRoleById(roleId){
     try{
+        if(!roleId){
+            throw new Error("Dati ID za role nije pronadjen");
+        }
         const response = await api.get(`${import.meta.env.VITE_API_BASE_URL}/roles/role/${roleId}`,{
             headers:getHeader()
         });
